refactor: announcements subfolder
