+++ .changeset/spicy-snakes-occur.md
@@ -1,0 +1,5 @@
+---
+'@procore-oss/backstage-plugin-announcements': patch
+---
+
+Components have been organized into three sub folders - announcements, categories, homeComponents.
