+++ plugins/announcements/src/components/Router.tsx
@@ -11,10 +11,10 @@ import {
   announcementViewRouteRef,
   categoriesListRouteRef,
 } from '../routes';
-import { AnnouncementsPage } from './AnnouncementsPage';
-import { AnnouncementPage } from './AnnouncementPage';
-import { CreateAnnouncementPage } from './CreateAnnouncementPage';
-import { EditAnnouncementPage } from './EditAnnouncementPage';
+import { AnnouncementsPage } from './announcements/AnnouncementsPage';
+import { AnnouncementPage } from './announcements/AnnouncementPage';
+import { CreateAnnouncementPage } from './announcements/CreateAnnouncementPage';
+import { EditAnnouncementPage } from './announcements/EditAnnouncementPage';
 import { CategoriesPage } from './categories';
 
 type RouterProps = {
