+++ plugins/announcements/src/components/announcements/AnnouncementPage/index.ts
