+++ plugins/announcements/src/components/announcements/AnnouncementSearchResultListItem/AnnouncementSearchResultListItem.tsx
