+++ plugins/announcements/src/components/announcements/AnnouncementSearchResultListItem/index.ts
