+++ plugins/announcements/src/components/announcements/AnnouncementsPage/AnnouncementsPage.test.tsx
@@ -5,9 +5,9 @@ import {
   renderInTestApp,
 } from '@backstage/test-utils';
 import { AnnouncementsPage } from './AnnouncementsPage';
-import { rootRouteRef } from '../../routes';
+import { rootRouteRef } from '../../../routes';
 import { permissionApiRef } from '@backstage/plugin-permission-react';
-import { announcementsApiRef } from '../../api';
+import { announcementsApiRef } from '../../../api';
 import { catalogApiRef, entityRouteRef } from '@backstage/plugin-catalog-react';
 import { fireEvent } from '@testing-library/react';
