+++ plugins/announcements/src/components/announcements/AnnouncementsPage/DeleteAnnouncementDialog.tsx
