+++ plugins/announcements/src/components/announcements/NewAnnouncementBanner/index.ts
