+++ plugins/announcements/src/components/announcements/index.ts
@@ -1,0 +1,7 @@
+export * from './AnnouncementForm';
+export * from './AnnouncementPage';
+export * from './AnnouncementsPage';
+export * from './EditAnnouncementPage';
+export * from './CreateAnnouncementPage';
+export * from './NewAnnouncementBanner';
+export * from './AnnouncementSearchResultListItem';
