+++ plugins/announcements/src/components/index.ts
@@ -1,0 +1,5 @@
+export { Router } from './Router';
+
+export * from './announcements';
+export * from './categories';
+export * from './homeComponents';
